Build 39

+++ section/product/product.js
@@ -1,15 +1,22 @@
 
 function product() {
     console.log("product function executed");
-
-    fetch('http://3.109.5.164/productData')
-        .then(response => {
-            if (!response.ok) {
-                throw new Error('Network response was not ok');
-            }
-            return response.json();
-        })
+    var loader = document.getElementById('loader');
+    loader.style.display = 'block';
+    fetch('http://65.0.168.11/productData')
+    .then(response => {
+        if (response.status === 404) {
+            loader.style.display = 'none';
+            alert("No data found.");
+            throw new Error('Data not found');
+        }
+        if (!response.ok) {
+            throw new Error('Network response was not ok');
+        }
+        return response.json();
+    })
         .then(data => {
+            loader.style.display = 'none';
             console.log(data);
             populateTable(data);
         })
@@ -23,6 +30,7 @@ function populateTable(data) {
     tbody.innerHTML = ''; // Clear existing rows
     var columnsToDisplay = ['name', 'rate'];
     var counter = 1;
+    var isAdmin = JSON.parse(localStorage.getItem('sessionData'))[0].usertype === 'Admin';
     data.forEach(function(item) {
         var row = tbody.insertRow();
         var cell = row.insertCell();
@@ -32,43 +40,36 @@ function populateTable(data) {
             cell.textContent = item[key];
         });
 
-         // Add Edit button~
-         var editCell = row.insertCell();
-         var editButton = document.createElement('button');
-         editButton.className = 'button edit-button';
-         var editLink = document.createElement('a');
-         editLink.href = '../product/update_product.html'; // Edit link destination
-         editLink.textContent = 'Edit';
-         editButton.appendChild(editLink);
-
-        editButton.addEventListener('click', function() {
-            editProduct(item); // Pass the user data to the edit function
-        });
-        editCell.appendChild(editButton);
+         // Add Edit button if user is admin
+         if (isAdmin) {
+            var editCell = row.insertCell();
+            var editButton = document.createElement('button');
+            editButton.className = 'button edit-button';
+            var editLink = document.createElement('a');
+            editLink.href = '../product/update_product.html'; // Edit link destination
+            editLink.textContent = 'Edit';
+            editButton.appendChild(editLink);
+            editButton.addEventListener('click', function() {
+                editProduct(item); // Pass the user data to the edit function
+            });
+            editCell.appendChild(editButton);
+        }
  
-         // Add Delete button
-         var deleteCell = row.insertCell();
-         var deleteButton = document.createElement('button');
-         deleteButton.className = 'button delete-button';
-         deleteButton.textContent = 'Delete';
-         deleteButton.addEventListener('click', function() {
-            deleteProduct(item.id); // Pass the user id to the delete function
-        });
-        deleteCell.appendChild(deleteButton);
+        // Add Delete button if user is admin
+        if (isAdmin) {
+            var deleteCell = row.insertCell();
+            var deleteButton = document.createElement('button');
+            deleteButton.className = 'button delete-button';
+            deleteButton.textContent = 'Delete';
+            deleteButton.addEventListener('click', function() {
+                deleteProduct(item.id); // Pass the user id to the delete function
+            });
+            deleteCell.appendChild(deleteButton);
+        }
     });
 }
 
 function editProduct(user) {
-    // Convert user data to JSON and encode it for URL
-    // var userData = encodeURIComponent(JSON.stringify(user));
-    // console.log("Jell")
-    // console.log(userData)
-    // console.log(user.id);
-    // document.getElementById("id1").value = user.id;
-
-    // Redirect to user_master.html with user data in query parameter
-    // window.location.href = "../user/User_Master.html"
-    // window.location.href = '../user/user_update.html?userData=' + '%7B"id"%3A31%2C"name"%3A"Deepali"%2C"address"%3A"nsk"%2C"mobile_no"%3A1234567890%2C"username"%3A"dee"%2C"password"%3A"asd"%2C"status"%3A"1"%2C"usertype"%3A"Admin"%7D';
     localStorage.setItem('userData', JSON.stringify(user));
      // Redirect to user_update.html
      window.location.href = '../product/update_product.html';
@@ -78,7 +79,7 @@ function editProduct(user) {
 
 function deleteProduct(userId) {
     // Perform delete operation based on userId
-    fetch('http://3.109.5.164/productData/deleteproductId/' + userId, {
+    fetch('http://65.0.168.11/productData/deleteproductId/' + userId, {
         method: 'DELETE'
     })
     .then(response => {
